I decided that I do want to save aadt fraction with each hour
This will make it possible to create a view that spots very high
aadt fractions.  Because views are idempotent and can only access
a single document, computing the fraction of annual daily traffic
is impossible because the view would have to access another doc in
the database (the aadt document for the particular year)

Also, decided that the couch_file code does not need to have grid
information because I am not saving the geometry, only the i,j cells,
which are now part of the task object.  Thus re-running this to save
aadt within each hour does not need to hit the postgis db to get the
geometry, and should go much faster

modified:   lib/couch_file.js
modified:   test/test_couch_files.js