get rid of periods that are greater than aadt
using R, figured out that there are some outliers
in the data which are at or above aadt.  This is
for hourly data.  Further testing in R showed that
one pass is probbly good enough to get the worst
offenders.

implemented that in the js, so that now it computes
aadt with all data, then goes back, drops the hours
with greater than aadt hourly values, and then
recomputes aadt.

Not done for hh or not_hh, at least, not yet.  assume
without proof that if aadt.n is fixed, the others
are corrected too, and that aadt.hh and aadt.not_hh
are similar to then n values.

This probably isn't the case, but I am being optimistic