instead of plus 1, do plus 2
I just ran a case that failed twice.  what a waste of time.  so to be
sure, add two to the number of groups variable, to cut the count
significantly.

Probably I should just start out with a target of 90 groups, because
that seems to be about right in most cases.