fixed a bug.  Sadly, everything I've done is suspect now.
When performing "get raft of grids", I made a mistake linking up the
lat lon with the couchdb fetch.  Not sure exactly where the bug lay,
but it is fixed now...I went through the creation of the matrix and
the merging of the data back onto the matrix.

Looking at the old code again, the problem was that it is not
guaranteed the way I was doing it that the index and the i_cell,
j_cell, lat, lon would end up the same as what they were in the
original grid passed in.