more bugs
the overlap code was ridiculously wrong.  not really a problem, but a
waste of cpu.  anyway, now that the fetch code is updated and I am
carrying geo_id, use that directly to identify overlaps.

works, checked with SF, day 7, month 10 cluster 6

Also a fix with the merge code (only use s.idx, not i_cell,j_cell too)