fixed the empty cluster bug
So the problem was that I was clustering grid cells based on the fact
that they *should* have data because the grid contained a detector
site.  This isn't good enough, because of course detectors come and go
over time, or have bad data, etc.

So instead, the fix is to first hit couchdb and load up the count of
the data in the month for each grid cell that *should* have data, and
then only keep those that actually *do*.

This is done inside of the month loop.

Then the clustering is done, and contains only sites with actual
data.  Again, this is per month.