fixed NCC one location cluster bug
if only one grid cell to cluster on, don't call cluster.  also, as it
was related, fix the assign.fraction code ---getting rid of the "buggy
version" carp/croak

made assign.fraction behave just like code in group.loop, with vars
changed as appropriate.

Running it proves that the memory leak is in the spTimer code, by the
way, as the RAM usage per thread stays fixed at 0.2%