basic test for get.grid.file, parseGridRecord
gets a detector grid file, and parses it.  Test passing means that you
have the right dimensions of the result.  I do not actually check any
values, etc.

But that's okay because the reason I wrote this test is that now the
data part of the stored object in couchdb is multiple lines
possibly (one per freeway) whereas the old code expected at most one
line.