fiddle do limit, and fix the returnval
can't remember exactly what we did over the phone, but this accomplishes
similar.

Basically, if we finish a cluster and get a zero result, don't return it
to caller, because that means we're going to bail out.  Instead return
1.

If the model loop exits with nothing to do every time, then it will
report -1, and so *then* we can report back to the outer script a zero
result to signify we're all done.